Group_No._14/FEMAI/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
python-socketio==5.8.0
python-engineio==4.7.1
requests==2.31.0
openai==1.3.0
speechrecognition==3.10.0
pyttsx3==2.90
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
python-dotenv==1.0.0
gunicorn==21.2.0

# WebRTC and Media Handling
aiortc==1.5.0
aiohttp==3.9.1
websockets==12.0
python-multipart==0.0.6

# Additional Dependencies for Enhanced Functionality
Pillow==10.0.1
numpy==1.24.3

# for vocie
Flask==2.3.0
flask-cors==4.0.0
pandas==2.0.0
numpy==1.24.0
scikit-learn==1.3.0
pickle-mixin==1.0.2
google-cloud-speech==2.21.0
pyaudio==0.2.13
wave==0.0.2
